This is the backgroung for the guy on the boulder.
If you think it looks a bit bizarre, you are completely right.
I was forced to photoshop the background from the margins of
the un-cut other image. It could use another run at it  -
but works at the moment mainly because the WB / L are all matched perfectly
(As they would be, they are from the same picture)