fix: Searches for climbs should consider parent areas, not descriptions
When searching for a climb, it is useful to be able to narrow the number
of results by including the parent areas' names.  open-tacos does not
include parent area names as a field when querying, so this did not work.
On the other hand, it does include descriptions, which results in many
spurious matches.

Fix this by searching on route names and area names instead of route
names and descriptions.

Fixes https://github.com/OpenBeta/open-tacos/issues/907